create config/reactions, fix lint
