+++ packages/atlas/src/views/viewer/VideoView/CommentsSection.tsx
@@ -13,9 +13,10 @@ import { Text } from '@/components/Text'
 import { Comment } from '@/components/_comments/Comment'
 import { CommentEditHistory } from '@/components/_comments/CommentEditHistory'
 import { CommentInput } from '@/components/_comments/CommentInput'
-import { REACTION_TYPE, ReactionChipProps, ReactionId } from '@/components/_comments/ReactionChip'
+import { ReactionChipProps } from '@/components/_comments/ReactionChip'
 import { Select } from '@/components/_inputs/Select'
 import { DialogModal } from '@/components/_overlays/DialogModal'
+import { REACTION_TYPE, ReactionId } from '@/config/reactions'
 import { absoluteRoutes } from '@/config/routes'
 import { COMMENTS_SORT_OPTIONS } from '@/config/sorting'
 import { useDisplaySignInDialog } from '@/hooks/useDisplaySignInDialog'
